backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
python-dotenv==1.1.0
pydantic==2.11.3
email-validator==2.1.0.post1 